requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-slugify==8.0.4
SQLAlchemy==2.0.36
text-unidecode==1.3
typing_extensions==4.12.2
Werkzeug==3.1.2
